+++ lib/components/Modal/index.d.ts
@@ -5,3 +5,4 @@ export * from './ModalHeader';
 export * from './modal.service';
 export * from './GlobalModal';
 export * from './modal.enum';
+export * from './modal.interfaces';