+++ lib/components/Modal/modal.enum.d.ts
@@ -1,9 +1,10 @@
-export declare enum MODALRESULT {
-    OK = "OK",
-    CANCEL = "CANCEL",
-    DELETE = "DELETE"
-}
 export declare enum MODALTYPE {
     BASIC = "BASIC",
     FORM = "FORM"
+}
+export declare enum MODALBUTTONTYPE {
+    OK = "OK",
+    CANCEL = "CANCEL",
+    RESET = "RESET",
+    DEFAULT = "DEFAULT"
 }