+++ lib/components/Modal/modal.service.d.ts
@@ -1,13 +1,13 @@
 import { IControls, IFormValues } from '../Form';
 import { ReactElement } from 'react';
+import { IModalButton } from './modal.interfaces';
 export interface IModalService {
     show(title: string, description: string | ReactElement, options?: IModalOptions): Promise<void | IFormValues>;
 }
 export interface IModalOptions {
     formControls?: IControls;
-    showOkButton?: boolean;
-    showCancelButton?: boolean;
     isDismissable?: boolean;
+    buttons?: Array<IModalButton>;
 }
 declare class ModalService implements IModalService {
     private container;